requirements.txt
python-telegram-bot==13.15
APScheduler==3.10.4
Flask==3.0.3
filelock==3.14.0
python-dateutil==2.9.0.post0
pytz==2024.1
uvicorn==0.30.1
orjson==3.10.6
